uhd: Fix rounding in ddc/duc rate calculation
The DDC and DUC convert the requested rate to an integer before
selecting a decimated / interpolated rate. This causes the selection to
select a lower rate than requested in some corner cases. The effect is
more pronounced when the input rate of the DDC or the output rate of the
DUC is very small.

+++ host/lib/rfnoc/ddc_block_ctrl_impl.cpp
@@ -245,7 +245,7 @@ private:
         return range;
     }
 
-    double set_output_rate(const int requested_rate, const size_t chan)
+    double set_output_rate(const double requested_rate, const size_t chan)
     {
         const double input_rate = get_arg<double>("input_rate");
         const size_t decim_rate =
